feedback button emoji on windows.
The error could be related to the usage of <>&#128557;</> within the button. It is not able to remove the child element from the DOM windows.  React might not be able to handle the html entities in  fragment components correctly within a button component, especially when the conditionals and entities are involved. Have wrapped it in a span element so it considers the span as a child element.