Avoid translating installed dependencies
This adds the `vendor`, `node_modules`, and `wordpress` directories to the exclude list for the `make-pot` command.

Because WordPress is installed as a Composer dependency, this was resulting in all WordPress Core strings being sent over to Azure for translation and including them in the langauge files only to never be used since Core does not use the module's text domain.